Move vue-i18n connection code to a separate file
