+++ src/Themes/default/scripts/light_portal/app_comments.js
@@ -1,24 +1,13 @@
 import { createApp } from 'vue';
 import { createPinia } from 'pinia';
-import { createI18n } from 'vue-i18n';
-import { VueShowdownPlugin, showdown } from 'vue-showdown';
+import { VueShowdown, showdown } from 'vue-showdown';
 import '@github/markdown-toolbar-element';
 import CommentList from '../../LightPortal/components/CommentList.vue';
-import Plurals from './dev/plurals';
+import i18n from './dev/i18n';
 
 const app = createApp(CommentList);
 
 app.use(createPinia());
-
-const plurals = new Plurals();
-
-const i18n = createI18n({
-  locale: vueGlobals.context.locale,
-  pluralizationRules: plurals.rules(),
-  messages: {
-    [vueGlobals.context.locale]: vueGlobals.txt,
-  },
-});
 
 app.use(i18n);
 
@@ -43,7 +32,7 @@ const bindings = Object.keys(classMap).map((key) => ({
 
 showdown.extension('bindings', bindings);
 
-app.use(VueShowdownPlugin, {
+app.use(VueShowdown, {
   flavor: 'github',
   options: {
     emoji: true,
