+++ src/Themes/default/scripts/light_portal/app_plugins.js
@@ -1,22 +1,11 @@
 import { createApp } from 'vue';
 import { createPinia } from 'pinia';
-import { createI18n } from 'vue-i18n';
 import PluginList from '../../LightPortal/components/PluginList.vue';
-import Plurals from './dev/plurals';
+import i18n from './dev/i18n';
 
 const app = createApp(PluginList);
 
 app.use(createPinia());
-
-const plurals = new Plurals();
-
-const i18n = createI18n({
-  locale: vueGlobals.context.locale,
-  pluralizationRules: plurals.rules(),
-  messages: {
-    [vueGlobals.context.locale]: vueGlobals.txt,
-  },
-});
 
 app.use(i18n);
